sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "valka"
version = "1.0.0"
description = "Python SDK for the Valka distributed task queue"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.25",
    "grpcio>=1.60",
    "protobuf>=4.25",
]

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.60",
    "mypy>=1.8",
    "ruff>=0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/valka"]
